requirements.txt
pandas
numpy
kaggle
pyyaml
pymysql
python-dotenv
scikit-learn==1.6.1
dill
xgboost
catboost
dagshub
mlflow

gunicorn
streamlit
shap
matplotlib
